+++ src/types/ArrayOrIterable.ts
@@ -1,10 +1,6 @@
-import type { ReadOnlyObject } from "./ReadOnlyObject";
-
 /**
- * Readonly ArrayLike or Iterable alias.
+ * ArrayLike or Iterable alias.
  *
  * @template Item Type of items in the ArrayLike or Iterable.
  */
-export type ArrayOrIterable<Item> =
-	| ReadOnlyObject<ArrayLike<Item>>
-	| Iterable<Item>;
+export type ArrayOrIterable<Item> = ArrayLike<Item> | Iterable<Item>;