+++ src/types/Grouper.ts
@@ -1,5 +1,3 @@
-import type { ReadOnlyObject } from "./ReadOnlyObject";
-
 /**
  * Grouper function.
  *
@@ -7,4 +5,4 @@ import type { ReadOnlyObject } from "./ReadOnlyObject";
  * @param item Item of the source array.
  * @returns Group name.
  */
-export type Grouper<Item> = (item: ReadOnlyObject<Item>) => string | number;
+export type Grouper<Item> = (item: Item) => string | number;